Conflicts:
	CMakeLists.txt
	src/cmake/modules/FindIlmBase.cmake
	src/include/CMakeLists.txt
	src/include/oslexec.h
	src/liboslexec/backendllvm.cpp
	src/liboslexec/backendllvm.h
	src/liboslexec/llvm_gen.cpp
	src/liboslexec/llvm_instance.cpp
	src/liboslexec/llvm_ops.cpp
	src/liboslexec/llvm_util.cpp
	src/liboslexec/llvmutil_test.cpp
	src/liboslexec/oslexec_pvt.h
	src/liboslexec/shadingsys.cpp

+++ src/liboslexec/typespec.cpp
@@ -30,9 +30,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <string>
 #include <cstdio>
 
-#include "OpenImageIO/strutil.h"
-#include "OpenImageIO/dassert.h"
-#include "OpenImageIO/thread.h"
+#include <OpenImageIO/strutil.h>
+#include <OpenImageIO/dassert.h>
+#include <OpenImageIO/thread.h>
 
 #include "oslexec_pvt.h"
 
@@ -101,7 +101,7 @@ equivalent (const StructSpec *a, const StructSpec *b)
     ASSERT (a && b);
     if (a->numfields() != b->numfields())
         return false;
-    for (size_t i = 0;  i < a->numfields();  ++i)
+    for (size_t i = 0;  i < (size_t)a->numfields();  ++i)
         if (! equivalent (a->field(i).type, b->field(i).type))
             return false;
     return true;
